requirements.txt
altgraph==0.17.4
asgiref==3.8.1
bidict==0.23.1
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==4.2.15
djangorestframework==3.15.2
dnspython==2.6.1
et-xmlfile==1.1.0
eventlet==0.36.1
Flask==3.0.3
Flask-SocketIO==5.3.6
fonttools==4.51.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
numpy==1.26.4
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pillow==10.3.0
psycopg2==2.9.9
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pymongo==4.7.1
pyparsing==3.1.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
PyQt5Designer==5.14.1
PySide2==5.15.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
pywin32-ctypes==0.2.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.31.0
shiboken2==5.15.2.1
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.2
wsproto==1.2.0
zipp==3.18.1
